+++ metrics_replications.go
@@ -3,12 +3,14 @@ package main
 import (
 	"encoding/json"
 	"strconv"
+	"time"
 
 	"github.com/go-kit/kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
 func (e *HarborExporter) collectReplicationsMetric(ch chan<- prometheus.Metric) bool {
+	start := time.Now()
 	type policiesMetrics []struct {
 		Id   float64
 		Name string
@@ -23,10 +25,17 @@ func (e *HarborExporter) collectReplicationsMetric(ch chan<- prometheus.Metric) 
 		// Extra fields omitted for maintainability: not relevant for current metrics
 	}
 
-	policiesBody, _ := e.request("/replication/policies")
 	var policiesData policiesMetrics
+	err := e.requestAll("/replication/policies", func(pageBody []byte) error {
+		var pageData policiesMetrics
+		if err := json.Unmarshal(pageBody, &pageData); err != nil {
+			return err
+		}
+		policiesData = append(policiesData, pageData...)
 
-	if err := json.Unmarshal(policiesBody, &policiesData); err != nil {
+		return nil
+	})
+	if err != nil {
 		level.Error(e.logger).Log("msg", "Error retrieving replication policies", "err", err.Error())
 		return false
 	}
@@ -66,5 +75,7 @@ func (e *HarborExporter) collectReplicationsMetric(ch chan<- prometheus.Metric) 
 			)
 		}
 	}
+
+	reportLatency(start, "replication_latency", ch)
 	return true
 }